Seperate hook into multiple file handling hooks
