+++ front/app/hooks/files/types.ts
@@ -2,11 +2,16 @@ import { UploadFile } from 'typings';
 
 import { FileType } from 'components/UI/FileUploader/FileDisplay';
 
-export type SyncFilesArguments = {
-  projectId?: string;
-  phaseId?: string;
-  projectFiles?: UploadFile[];
-  phaseFiles?: FileType[];
+export type SyncProjectFilesArguments = {
+  projectId: string;
+  projectFiles: UploadFile[];
+  filesToRemove: any[];
+  fileOrdering: Record<string, number | undefined>;
+};
+
+export type SyncPhaseFilesArguments = {
+  phaseId: string;
+  phaseFiles: FileType[];
   filesToRemove: any[];
   fileOrdering: Record<string, number | undefined>;
 };
